services/ingestion/pyproject.toml
[project]
name = "ingestion"
version = "0.1.0"
description = "Ingestion Service for OctaneBrew"
readme = "README.md"
authors = [
    { name = "shubh305", email = "shubham.305singh@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "aiokafka>=0.13.0",
    "elasticsearch>=9.2.1",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "langchain>=1.2.7",
    "langchain-experimental>=0.4.1",
    "langchain-text-splitters>=1.1.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "beautifulsoup4>=4.12.3",
    "asyncpg>=0.29.0",
    "kafka-python>=2.0.2",
    "tenacity>=8.2.3",
    "fastapi-limiter>=0.1.6",
    "redis>=5.0.1",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "opentelemetry-instrumentation-kafka-python>=0.43b0",
    "opentelemetry-instrumentation-httpx>=0.43b0",
    "pydantic-settings>=2.12.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "tiktoken>=0.7.0",
]

[project.scripts]
ingestion = "ingestion:main"

[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"
